switch lidar mode command
